@@ -13,24 +13,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.10",
-    "@angular/cdk": "^16.2.9",
-    "@angular/common": "^16.2.10",
-    "@angular/compiler": "^16.2.10",
-    "@angular/core": "^16.2.10",
-    "@angular/forms": "^16.2.10",
-    "@angular/language-service": "^16.2.10",
-    "@angular/material": "^16.2.9",
-    "@angular/platform-browser": "^16.2.10",
-    "@angular/platform-browser-dynamic": "^16.2.10",
-    "@angular/router": "^16.2.10",
-    "@coreui/angular": "~4.5.27",
-    "@coreui/angular-chartjs": "~4.5.27",
+    "@angular/animations": "^17.3.3",
+    "@angular/cdk": "^17.3.3",
+    "@angular/common": "^17.3.3",
+    "@angular/compiler": "^17.3.3",
+    "@angular/core": "^17.3.3",
+    "@angular/forms": "^17.3.3",
+    "@angular/language-service": "^17.3.3",
+    "@angular/material": "^17.3.1",
+    "@angular/platform-browser": "^17.3.3",
+    "@angular/platform-browser-dynamic": "^17.3.3",
+    "@angular/router": "^17.3.3",
+    "@coreui/angular": "~4.7.15",
+    "@coreui/angular-chartjs": "~4.7.15",
     "@coreui/chartjs": "^3.1.2",
     "@coreui/coreui": "~4.2.6",
     "@coreui/icons": "^3.0.1",
-    "@coreui/icons-angular": "~4.5.27",
+    "@coreui/icons-angular": "~4.7.15",
     "@coreui/utils": "^2.0.2",
+    "@syncfusion/ej2-angular-filemanager": "^25.1.39",
+    "@syncfusion/ej2-angular-grids": "^25.1.39",
+    "@syncfusion/ej2-angular-kanban": "^25.1.35",
+    "@syncfusion/ej2-angular-schedule": "^25.1.39",
+    "@syncfusion/ej2-material-theme": "~25.1.35",
     "bootstrap": "^5.3.3",
     "bootstrap-icons": "^1.11.3",
     "chart.js": "^3.9.1",
@@ -40,13 +45,13 @@
     "rxjs": "~7.8.1",
     "sweetalert2": "^11.10.7",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.1"
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.7",
-    "@angular/cli": "^16.2.7",
-    "@angular/compiler-cli": "^16.2.10",
-    "@angular/localize": "^16.2.10",
+    "@angular-devkit/build-angular": "^17.3.3",
+    "@angular/cli": "^17.3.3",
+    "@angular/compiler-cli": "^17.3.3",
+    "@angular/localize": "^17.3.3",
     "@types/jasmine": "^5.1.1",
     "@types/lodash-es": "^4.17.10",
     "@types/node": "^18.18.6",
@@ -56,7 +61,7 @@
     "karma-coverage": "^2.2.1",
     "karma-jasmine": "^5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "typescript": "~4.9.5"
+    "typescript": "~5.4.4"
   },
   "engines": {
     "node": "^16.14.0 || ^18.10.0",
