@@ -3,16 +3,11 @@ import { INavData } from '@coreui/angular';
 export const navItems: INavData[] = [
   {
     title: true,
-    name: 'Clientes y Leads'
+    name: 'Clientes y Contactos'
   },
-  // {
-  //   name: 'Contactos',
-  //   url: '/contacts/leads',
-  //   iconComponent: { name: 'cil-user-follow' },
-  // },
   {
     name: 'Clientes',
-    url: '/contact/customer',
+    url: '/contact/customer-list',
     iconComponent: { name: 'cil-group' },
   },
   {
@@ -21,12 +16,12 @@ export const navItems: INavData[] = [
   },
   {
     name: 'Cotizaciones',
-    url: '/sales/quote',
+    url: '/sales/quote-list',
     iconComponent: { name: 'cil-calculator' },
   },
   {
     name: 'Ventas',
-    url: '/sales/sale',
+    url: '/sales/sale-list',
     iconComponent: { name: 'cil-cart' },
   },
   {
@@ -35,12 +30,17 @@ export const navItems: INavData[] = [
   },
   {
     name: 'Proyectos',
-    url: '/projects/project',
+    url: '/projects/project-list',
     iconComponent: { name: 'cil-briefcase' },
   },
   {
     name: 'Tareas',
-    url: '/projects/tasks',
+    url: '/projects/task-list',
     iconComponent: { name: 'cil-bolt' },
-  }
+  },
+  // {
+  //   name: 'Pizarra Kanban',
+  //   url: '/projects/task-kanban',
+  //   iconComponent: { name: 'cil-clipboard' },
+  // }
 ];
