@@ -1,7 +1,7 @@
 import { Component, Inject } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { Customer } from 'src/app/modules/contacts/models/Customer';
+import { Customer, CustomerData, CustomerStatus } from 'src/app/modules/contacts/models/Customer';
 import { CustomerService } from '../../services/customer.service';
 
 @Component({
@@ -11,7 +11,29 @@ import { CustomerService } from '../../services/customer.service';
 })
 export class CustomerModalComponent {
 
-  statusList = ['New', 'In Progress', 'Done'];
+  statusTypes : CustomerStatus[] = [
+    {
+      id: 1,
+      code: 'NEW',
+      name: 'Nuevo'
+    },
+    {
+      id: 2,
+      code: 'LEAD',
+      name: 'Lead'
+    },
+    {
+      id: 3,
+      code: 'CUSTOMER',
+      name: 'Cliente'
+    },
+    {
+      id: 4,
+      code: 'NOT_INTERESTED',
+      name: 'No interesado'
+    }
+  ]
+
 
   action: string = "";
 
@@ -20,33 +42,25 @@ export class CustomerModalComponent {
     lastName: [''],
     rut: [''],
     email: [''],
+    contactNumber: [''],
+    address: [''],
     city: [''],
-    contactPlatform: [''],
-    budget: [0],
-    projectDescription: [''],
-    expectedDate: [''],
-    deliveryStatus: [''],
-    workLine: [''],
+    domain: [''],
+    type: [''],
     inCharge: [''],
-    status: [''],
+    contactPlatform: [''],
+    description: [''],
+    workline: [''],
+    projects: [''],
+    createdAt: [''],
+    createdBy: [''],
+    lastModifiedAt: [''],
+    lastModifiedBy: [''],
+    status: ['']
   });
 
-  customer: Customer = {
-    id: 0,
-    firstName: '',
-    lastName: '',
-    rut: '',
-    email: '',
-    city: '',
-    contactPlatform: '',
-    budget: 0,
-    projectDescription: '',
-    expectedDate: '',
-    deliveryStatus: '',
-    workLine: '',
-    inCharge: '',
-    status: ''
-  }
+
+  customer: Customer = {} as Customer
 
   constructor(
     @Inject(MAT_DIALOG_DATA) public data: any,
@@ -69,38 +83,45 @@ export class CustomerModalComponent {
       .subscribe((response: Customer) => {
         this.customer = response;
         this.clientForm.setValue({
-          firstName: this.customer.firstName,
-          lastName: this.customer.lastName,
-          rut: this.customer.rut,
-          email: this.customer.email,
-          city: this.customer.city,
-          contactPlatform: this.customer.contactPlatform,
-          budget: this.customer.budget,
-          projectDescription: this.customer.projectDescription,
-          expectedDate: this.customer.expectedDate,
-          deliveryStatus: this.customer.deliveryStatus,
-          workLine: this.customer.workLine,
-          inCharge: this.customer.inCharge,
-          status: this.customer.status
+          firstName: this.customer.firstName!,
+          lastName: this.customer.lastName!,
+          rut: this.customer.rut!,
+          email: this.customer.email!,
+          contactNumber: this.customer.contactNumber!,
+          contactPlatform: this.customer.contactPlatform!,
+          workline: this.customer.workline!,
+          domain: this.customer.domain!,
+          address: this.customer.address!,
+          city: this.customer.city!,
+          description: this.customer.description!,
+          inCharge: this.customer.inCharge!,
+          type: String(this.customer.type!),
+          status: this.clientForm.value.status!,
+          projects: this.customer.projects!,
+          createdAt: this.customer.createdAt!,
+          createdBy: this.customer.createdBy!,
+          lastModifiedAt: this.customer.lastModifiedAt!,
+          lastModifiedBy: this.customer.lastModifiedBy!
         })
       })
   }
 
   createOrUpdate() {
-    const request: Customer = {
+    const request: CustomerData = {
       id: this.data.id,
       firstName: this.clientForm.value.firstName!,
       lastName: this.clientForm.value.lastName!,
       rut: this.clientForm.value.rut!,
       email: this.clientForm.value.email!,
+      contactNumber: this.clientForm.value.contactNumber!,
+      workline: this.clientForm.value.workline!,
+      domain: this.clientForm.value.domain!,
+      address: this.clientForm.value.address!,
       city: this.clientForm.value.city!,
       contactPlatform: this.clientForm.value.contactPlatform!,
-      budget: this.clientForm.value.budget!,
-      projectDescription: this.clientForm.value.projectDescription!,
-      expectedDate: this.clientForm.value.expectedDate!,
-      deliveryStatus: this.clientForm.value.deliveryStatus!,
-      workLine: this.clientForm.value.workLine!,
+      description: this.clientForm.value.description!,
       inCharge: this.clientForm.value.inCharge!,
+      type: Number(this.clientForm.value.type!),
       status: this.clientForm.value.status!
     };
 
