@@ -1,18 +1,53 @@
 export interface Task {
   id: number;
-  title: string;
-  shortDesc: string;
-  deliveryStatus: string;
-  inCharge: string;
-  expectedDate: string;
   projectId: number;
+  title: string;
+  resume: string;
+  description: string;
+  objectives: string;
+  taskStatus: string;
+  inCharge: string;
+  priority: string;
+  comments: string;
+  startDate: string;
+  endDate: string;
+  tags: string;
+  files: string[];
+  createdAt: string;
+  createdBy: string;
+  lastModifiedAt: string;
+  lastModifiedBy: string;
 }
+
 export interface TaskData {
   id: number;
-  title: string;
-  shortDesc: string;
-  deliveryStatus: string;
-  inCharge: string;
-  expectedDate: string;
   projectId: number;
+  title: string;
+  resume: string;
+  taskStatus: string;
+  inCharge: string;
+  priority: string;
+  startDate: string;
+  endDate: string;
+}
+
+export interface TaskDetail {
+  id: number;
+  projectId: number;
+  title: string;
+  resume: string;
+  description: string;
+  objectives: string;
+  taskStatus: string;
+  inCharge: string;
+  priority: string;
+  comments: string;
+  startDate: string;
+  endDate: string;
+  tags: string;
+  files: string[];
+  createdAt: string;
+  createdBy: string;
+  lastModifiedAt: string;
+  lastModifiedBy: string;
 }
