@@ -1,7 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { TasksComponent } from './views/tasks/tasks.component';
-import { ProjectComponent } from './views/project/project.component';
+import { TaskListComponent } from './views/task/task-list/task-list.component';
+import { TaskDetailComponent } from './views/task/task-detail/task-detail.component';
+import { ProjectListComponent } from './views/project/project-list/project-list.component';
+import { ProjectDetailComponent } from './views/project/project-detail/project-detail.component';
+import { TaskKanbanComponent } from './views/task/task-kanban/task-kanban.component';
 
 const routes: Routes = [
   {
@@ -16,17 +19,38 @@ const routes: Routes = [
         redirectTo: 'tasks'
       },
       {
-        path: 'project',
-        component: ProjectComponent,
+        path: 'project-list',
+        component: ProjectListComponent,
         data: {
           title: 'Projects',
         },
       },
       {
-        path: 'tasks',
-        component: TasksComponent,
+        path: 'project-detail',
+        component: ProjectDetailComponent,
+        data: {
+          title: 'Project Details',
+        },
+      },
+      {
+        path: 'task-list',
+        component: TaskListComponent,
         data: {
           title: 'Tasks',
+        },
+      },
+      {
+        path: 'task-detail',
+        component: TaskDetailComponent,
+        data: {
+          title: 'Task Detail',
+        },
+      },
+      {
+        path: 'task-kanban',
+        component: TaskKanbanComponent,
+        data: {
+          title: 'Task Kanban',
         },
       },
     ]
