@@ -1,38 +1,41 @@
 export interface Quote {
   id: number;
-  clientId: string;
+  clientId: number;
+  clientName: string;
   title: string;
-  shortDesc: string;
-  longDesc: string;
-  objectives: string;
-  notes: string;
+  resume: string;
+  description: string;
   startDate: string;
   endDate: string;
   domain: string;
-  briefUrl: string;
-  details: string;
+  quoteStatus: string;
   estimatedCost: number;
   estimatedTime: string;
-}
-
-export interface QuoteData {
-  id: number;
-  clientId: string;
-  title: string;
-  shortDesc: string;
-  longDesc: string;
-  objectives: string;
-  notes: string;
-  startDate: string;
-  endDate: string;
-  domain: string;
-  briefUrl: string;
-  details: string;
-  estimatedCost: number;
-  estimatedTime: string;
+  inCharge: string;
   createdAt: string;
   createdBy: string;
   lastModifiedAt: string;
   lastModifiedBy: string;
-  status: string;
+}
+
+export interface QuoteData {
+  id: number;
+  clientId: number;
+  clientName: string;
+  title: string;
+  resume: string;
+  description: string;
+  startDate: string;
+  endDate: string;
+  domain: string;
+  quoteStatus: string;
+  estimatedCost: number;
+  estimatedTime: string;
+  inCharge: string;
+}
+
+export interface QuoteStatus {
+  id: number;
+  code: string;
+  name: string;
 }
