@@ -1,44 +1,51 @@
 export interface Sale {
   id: number;
-  clientId: string;
-  quoteId: string;
+  quoteId: number | null;
+  clientId: number;
+  clientName: string;
   title: string;
-  shortDesc: string;
-  longDesc: string;
-  objectives: string;
-  notes: string;
+  resume: string;
+  description: string;
   startDate: string;
   endDate: string;
   domain: string;
   briefUrl: string;
-  details: string;
-  estimatedCost: number;
-  estimatedTime: string;
+  cost: number;
+  duration: string;
   inCharge: string;
-  projectId: string;
-}
-
-export interface SaleData {
-  id: number;
-  clientId: string;
-  quoteId: string;
-  title: string;
-  shortDesc: string;
-  longDesc: string;
-  objectives: string;
-  notes: string;
-  startDate: string;
-  endDate: string;
-  domain: string;
-  briefUrl: string;
-  details: string;
-  estimatedCost: number;
-  estimatedTime: string;
-  inCharge: string;
-  projectId: string;
+  projectId: number;
+  projectStatus: string;
+  paymentType: string;
+  bruteCost: number;
+  iva: number;
+  total: number;
+  paymentStatus: string;
   createdAt: string;
   createdBy: string;
   lastModifiedAt: string;
   lastModifiedBy: string;
-  status: string;
+}
+
+export interface SaleData {
+  id: number;
+  clientName: string;
+  title: string;
+  resume: string;
+  startDate: string;
+  endDate: string;
+  domain: string;
+  briefUrl: string;
+  cost: number;
+  duration: string;
+  inCharge: string;
+  total: number;
+  paymentStatus: string;
+  createdAt: string;
+  createdBy: string;
+}
+
+export interface SaleStatus {
+  id: number;
+  code: string;
+  name: string;
 }
