@@ -1,1 +1,7 @@
-export const environment = {};
+export const environment = {
+  production: false,
+  clientMicroserviceUrl: "http://localhost:8081/api",
+  saleMicroserviceUrl: "http://localhost:8082/api",
+  projectMicroserviceUrl: "http://localhost:8083/api",
+  ej2LicenseKey: "Ngo9BigBOggjHTQxAR8/V1NBaF5cXmpCf1FpRmJGdld5fUVHYVZUTXxaS00DNHVRdkdnWXtfcHZVRGVfV0BxVks="
+};
