@@ -1,5 +1,7 @@
 export const environment = {
-  clientMicroserviceUrl: "http://localhost:8081/api",
-  saleMicroserviceUrl: "http://localhost:8082/api",
-  projectMicroserviceUrl: "http://localhost:8083/api"
+  production: true,
+  clientMicroserviceUrl: "http://192.168.49.2:30001/api",
+  saleMicroserviceUrl: "http://192.168.49.2:30002/api",
+  projectMicroserviceUrl: "http://192.168.49.2:30003/api",
+  ej2LicenseKey: "Ngo9BigBOggjHTQxAR8/V1NBaF5cXmpCf1FpRmJGdld5fUVHYVZUTXxaS00DNHVRdkdnWXtfcHZVRGVfV0BxVks="
 };
